refactor(server): TemplateDoc ドメインモデル欠乏症を解消し LiveTemplateDoc を導入
CRDT マージ・shape 検証・サイズ制約・差分計算を usecase 層 (shape.ts) から
model 層 (LiveTemplateDoc class) に移動し、usecase を薄いオーケストレーションに変える。

- constraints.ts: SYNC_LIMITS と許可キー定数群を model 層に集約
- live.ts: LiveTemplateDoc class (factory / 入力ガード / CRDT 操作 / 検証 / 差分計算 / 永続変換)
- templateDoc.ts: TemplateDoc interface + namespace を barrel から分離 (循環依存解消)
- index.ts: barrel re-export のみに簡素化
- syncTemplateDoc usecase: Y.Doc 直接操作 → LiveTemplateDoc API 委譲
- shape.ts / shape.test.ts: 削除 (ロジックは live.ts + constraints.ts に移植)
